// project/src/types/database.types.ts
export interface Database {
  public: {
    Tables: {
// ...
      contacts: {
        Row: {
          id: string
          created_at: string
          name: string
          email: string
          subject: string
          message: string
          read: boolean
          user_id: string
        }
        Insert: {
          id?: string
          created_at?: string
          name: string
          email: string
          subject: string
          message: string
          read?: boolean
          user_id: string
        }
        Update: {
          id?: string
          created_at?: string
          name?: string
          email?: string
          subject?: string
          message?: string
          read?: boolean
          user_id?: string
        }
      }
      profiles: {
        Row: {
          full_name: string | undefined
          id: string
          created_at: string
          updated_at: string
          name: string
          title: string
          bio: string
          avatar_url: string | null
          location: string | null
          email: string | null
          phone: string | null
          linkedin_url: string | null
          github_url: string | null
          twitter_url: string | null
        }
        Insert: {
          id: string
          created_at?: string
          updated_at?: string
          name: string
          title: string
          bio: string
          avatar_url?: string | null
          location?: string | null
          email?: string | null
          phone?: string | null
          linkedin_url?: string | null
          github_url?: string | null
          twitter_url?: string | null
        }
        Update: {
          id?: string
          created_at?: string
          updated_at?: string
          name?: string
          title?: string
          bio?: string
          avatar_url?: string | null
          location?: string | null
          email?: string | null
          phone?: string | null
          linkedin_url?: string | null
          github_url?: string | null
          twitter_url?: string | null
        }
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
  }
}